Split app-render functions into separate files
