+++ packages/next/src/server/app-render/flight-response-ref.tsx
@@ -1,0 +1,3 @@
+export interface FlightResponseRef {
+  current: Promise<JSX.Element> | null
+}
