+++ packages/next/src/server/app-render/get-server-css-for-entries.tsx
@@ -1,0 +1,18 @@
+import { ClientCSSReferenceManifest } from '../../build/webpack/plugins/flight-manifest-plugin'
+
+export function getServerCSSForEntries(
+  serverCSSManifest: ClientCSSReferenceManifest,
+  entries: string[]
+) {
+  const css = []
+  for (const entry of entries) {
+    const entryName = entry.replace(/\.[^.]+$/, '')
+    if (
+      serverCSSManifest.cssModules &&
+      serverCSSManifest.cssModules[entryName]
+    ) {
+      css.push(...serverCSSManifest.cssModules[entryName])
+    }
+  }
+  return css
+}
