+++ packages/next/src/server/app-render/index.tsx
@@ -1,12 +1,6 @@
 import type { IncomingMessage, ServerResponse } from 'http'
-import type { NextFontManifest } from '../../build/webpack/plugins/next-font-manifest-plugin'
-import {
-  ClientCSSReferenceManifest,
-  ClientReferenceManifest,
-} from '../../build/webpack/plugins/flight-manifest-plugin'
 import type {
   ChildProp,
-  DynamicParamTypes,
   DynamicParamTypesShort,
   FlightData,
   FlightDataPath,
@@ -20,48 +14,35 @@ import type { RequestAsyncStorage } from '../../client/components/request-async-
 import type { MetadataItems } from '../../lib/metadata/resolve-metadata'
 
 // Import builtin react directly to avoid require cache conflicts
-import React, { use } from 'next/dist/compiled/react'
+import React from 'next/dist/compiled/react'
 import { NotFound as DefaultNotFound } from '../../client/components/error'
 
 // this needs to be required lazily so that `next-server` can set
 // the env before we require
-import ReactDOMServer from 'next/dist/compiled/react-dom/server.browser'
 import { ParsedUrlQuery } from 'querystring'
 import { NextParsedUrlQuery } from '../request-meta'
 import RenderResult, { type RenderResultMetadata } from '../render-result'
 import {
-  readableStreamTee,
-  encodeText,
-  decodeText,
   renderToInitialStream,
   createBufferedTransformStream,
   continueFromInitialStream,
-  streamToString,
   streamToBufferedResult,
 } from '../node-web-streams-helper'
-import { ESCAPE_REGEX, htmlEscapeJsonString } from '../htmlescape'
 import { matchSegment } from '../../client/components/match-segments'
 import { ServerInsertedHTMLContext } from '../../shared/lib/server-inserted-html'
 import { stripInternalQueries } from '../internal-utils'
-import { DYNAMIC_ERROR_CODE } from '../../client/components/hooks-server-context'
 import { HeadManagerContext } from '../../shared/lib/head-manager-context'
-import stringHash from 'next/dist/compiled/string-hash'
 import {
   ACTION,
   NEXT_ROUTER_PREFETCH,
   NEXT_ROUTER_STATE_TREE,
   RSC,
-  RSC_CONTENT_TYPE_HEADER,
 } from '../../client/components/app-router-headers'
-import { formatServerError } from '../../lib/format-server-error'
 import { MetadataTree } from '../../lib/metadata/metadata'
 import { RequestAsyncStorageWrapper } from '../async-storage/request-async-storage-wrapper'
 import { StaticGenerationAsyncStorageWrapper } from '../async-storage/static-generation-async-storage-wrapper'
 import { collectMetadata } from '../../lib/metadata/resolve-metadata'
-import {
-  getClientReferenceModuleKey,
-  isClientReference,
-} from '../../lib/client-reference'
+import { isClientReference } from '../../lib/client-reference'
 import { getLayoutOrPageModule, LoaderTree } from '../lib/app-dir-module'
 import { warnOnce } from '../../shared/lib/utils/warn-once'
 import { isNotFoundError } from '../../client/components/not-found'
@@ -69,561 +50,25 @@ import {
   getURLFromRedirectError,
   isRedirectError,
 } from '../../client/components/redirect'
-import { NEXT_DYNAMIC_NO_SSR_CODE } from '../../shared/lib/lazy-dynamic/no-ssr-error'
 import { patchFetch } from '../lib/patch-fetch'
 import { AppRenderSpan } from '../lib/trace/constants'
 import { getTracer } from '../lib/trace/tracer'
-import { flightRouterStateSchema } from './types'
+import { interopDefault } from './interop-default'
+import { preloadComponent } from './preload-component'
+import { FlightRenderResult } from './flight-render-result'
+import { createErrorHandler } from './create-error-handler'
+import { createServerComponentRenderer } from './create-server-components-renderer'
+import { getShortDynamicParamType } from './get-short-dynamic-param-type'
+import { getSegmentParam } from './get-segment-param'
+import { getCssInlinedLinkTags } from './get-css-inlined-link-tags'
+import { getServerCSSForEntries } from './get-server-css-for-entries'
+import { getPreloadedFontFilesInlineLinkTags } from './get-preloaded-font-files-inline-link-tags'
+import { getScriptNonceFromHeader } from './get-script-nonce-from-header'
+import { renderToString } from './render-to-string'
+import { parseAndValidateFlightRouterState } from './parse-and-validate-flight-router-state'
+import { validateURL } from './validate-url'
 
-const isEdgeRuntime = process.env.NEXT_RUNTIME === 'edge'
-
-function preloadComponent(Component: any, props: any) {
-  const prev = console.error
-  // Hide invalid hook call warning when calling component
-  console.error = function (msg) {
-    if (msg.startsWith('Warning: Invalid hook call.')) {
-      // ignore
-    } else {
-      // @ts-expect-error argument is defined
-      prev.apply(console, arguments)
-    }
-  }
-  try {
-    let result = Component(props)
-    if (result && typeof result.then === 'function') {
-      // Catch promise rejections to prevent unhandledRejection errors
-      result.then(
-        () => {},
-        () => {}
-      )
-    }
-    return function () {
-      // We know what this component will render already.
-      return result
-    }
-  } catch (x) {
-    // something suspended or errored, try again later
-  } finally {
-    console.error = prev
-  }
-  return Component
-}
-
-/**
- * Flight Response is always set to RSC_CONTENT_TYPE_HEADER to ensure it does not get interpreted as HTML.
- */
-class FlightRenderResult extends RenderResult {
-  constructor(response: string | ReadableStream<Uint8Array>) {
-    super(response, { contentType: RSC_CONTENT_TYPE_HEADER })
-  }
-}
-
-/**
- * Action Response is set to application/json for now, but could be changed in the future.
- */
-class ActionRenderResult extends RenderResult {
-  constructor(response: string) {
-    super(response, { contentType: 'application/json' })
-  }
-}
-
-/**
- * Interop between "export default" and "module.exports".
- */
-function interopDefault(mod: any) {
-  return mod.default || mod
-}
-
-// tolerate dynamic server errors during prerendering so console
-// isn't spammed with unactionable errors
-/**
- * Create error handler for renderers.
- */
-function createErrorHandler({
-  /**
-   * Used for debugging
-   */
-  _source,
-  dev,
-  isNextExport,
-  errorLogger,
-  capturedErrors,
-  allCapturedErrors,
-}: {
-  _source: string
-  dev?: boolean
-  isNextExport?: boolean
-  errorLogger?: (err: any) => Promise<void>
-  capturedErrors: Error[]
-  allCapturedErrors?: Error[]
-}) {
-  return (err: any): string => {
-    if (allCapturedErrors) allCapturedErrors.push(err)
-
-    if (
-      err &&
-      (err.digest === DYNAMIC_ERROR_CODE ||
-        isNotFoundError(err) ||
-        err.digest === NEXT_DYNAMIC_NO_SSR_CODE ||
-        isRedirectError(err))
-    ) {
-      return err.digest
-    }
-
-    // Format server errors in development to add more helpful error messages
-    if (dev) {
-      formatServerError(err)
-    }
-    // Used for debugging error source
-    // console.error(_source, err)
-
-    // Don't log the suppressed error during export
-    if (
-      !(
-        isNextExport &&
-        err?.message?.includes(
-          'The specific message is omitted in production builds to avoid leaking sensitive details.'
-        )
-      )
-    ) {
-      if (errorLogger) {
-        errorLogger(err).catch(() => {})
-      } else {
-        // The error logger is currently not provided in the edge runtime.
-        // Use `log-app-dir-error` instead.
-        // It won't log the source code, but the error will be more useful.
-        if (process.env.NODE_ENV !== 'production') {
-          const { logAppDirError } =
-            require('../dev/log-app-dir-error') as typeof import('../dev/log-app-dir-error')
-          logAppDirError(err)
-        }
-        if (process.env.NODE_ENV === 'production') {
-          console.error(err)
-        }
-      }
-    }
-
-    capturedErrors.push(err)
-    // TODO-APP: look at using webcrypto instead. Requires a promise to be awaited.
-    return stringHash(err.message + err.stack + (err.digest || '')).toString()
-  }
-}
-
-interface FlightResponseRef {
-  current: Promise<JSX.Element> | null
-}
-
-/**
- * Render Flight stream.
- * This is only used for renderToHTML, the Flight response does not need additional wrappers.
- */
-function useFlightResponse(
-  writable: WritableStream<Uint8Array>,
-  req: ReadableStream<Uint8Array>,
-  clientReferenceManifest: ClientReferenceManifest,
-  rscChunks: Uint8Array[],
-  flightResponseRef: FlightResponseRef,
-  nonce?: string
-): Promise<JSX.Element> {
-  if (flightResponseRef.current !== null) {
-    return flightResponseRef.current
-  }
-  const {
-    createFromReadableStream,
-  } = require('next/dist/compiled/react-server-dom-webpack/client.edge')
-
-  const [renderStream, forwardStream] = readableStreamTee(req)
-  const res = createFromReadableStream(renderStream, {
-    moduleMap: isEdgeRuntime
-      ? clientReferenceManifest.edgeSSRModuleMapping
-      : clientReferenceManifest.ssrModuleMapping,
-  })
-  flightResponseRef.current = res
-
-  let bootstrapped = false
-  // We only attach CSS chunks to the inlined data.
-  const forwardReader = forwardStream.getReader()
-  const writer = writable.getWriter()
-  const startScriptTag = nonce
-    ? `<script nonce=${JSON.stringify(nonce)}>`
-    : '<script>'
-  const textDecoder = new TextDecoder()
-
-  function read() {
-    forwardReader.read().then(({ done, value }) => {
-      if (value) {
-        rscChunks.push(value)
-      }
-
-      if (!bootstrapped) {
-        bootstrapped = true
-        writer.write(
-          encodeText(
-            `${startScriptTag}(self.__next_f=self.__next_f||[]).push(${htmlEscapeJsonString(
-              JSON.stringify([0])
-            )})</script>`
-          )
-        )
-      }
-      if (done) {
-        flightResponseRef.current = null
-        writer.close()
-      } else {
-        const responsePartial = decodeText(value, textDecoder)
-        const scripts = `${startScriptTag}self.__next_f.push(${htmlEscapeJsonString(
-          JSON.stringify([1, responsePartial])
-        )})</script>`
-
-        writer.write(encodeText(scripts))
-        read()
-      }
-    })
-  }
-  read()
-
-  return res
-}
-
-/**
- * Create a component that renders the Flight stream.
- * This is only used for renderToHTML, the Flight response does not need additional wrappers.
- */
-function createServerComponentRenderer<Props>(
-  ComponentToRender: (props: Props) => any,
-  ComponentMod: {
-    renderToReadableStream: any
-    __next_app_webpack_require__?: any
-  },
-  {
-    transformStream,
-    clientReferenceManifest,
-    serverContexts,
-    rscChunks,
-  }: {
-    transformStream: TransformStream<Uint8Array, Uint8Array>
-    clientReferenceManifest: NonNullable<RenderOpts['clientReferenceManifest']>
-    serverContexts: Array<
-      [ServerContextName: string, JSONValue: Object | number | string]
-    >
-    rscChunks: Uint8Array[]
-  },
-  serverComponentsErrorHandler: ReturnType<typeof createErrorHandler>,
-  nonce?: string
-): (props: Props) => JSX.Element {
-  // We need to expose the `__webpack_require__` API globally for
-  // react-server-dom-webpack. This is a hack until we find a better way.
-  if (ComponentMod.__next_app_webpack_require__) {
-    // @ts-ignore
-    globalThis.__next_require__ = ComponentMod.__next_app_webpack_require__
-
-    // @ts-ignore
-    globalThis.__next_chunk_load__ = () => Promise.resolve()
-  }
-
-  let RSCStream: ReadableStream<Uint8Array>
-  const createRSCStream = (props: Props) => {
-    if (!RSCStream) {
-      RSCStream = ComponentMod.renderToReadableStream(
-        <ComponentToRender {...(props as any)} />,
-        clientReferenceManifest.clientModules,
-        {
-          context: serverContexts,
-          onError: serverComponentsErrorHandler,
-        }
-      )
-    }
-    return RSCStream
-  }
-
-  const flightResponseRef: FlightResponseRef = { current: null }
-
-  const writable = transformStream.writable
-  return function ServerComponentWrapper(props: Props): JSX.Element {
-    const reqStream = createRSCStream(props)
-    const response = useFlightResponse(
-      writable,
-      reqStream,
-      clientReferenceManifest,
-      rscChunks,
-      flightResponseRef,
-      nonce
-    )
-    return use(response)
-  }
-}
-
-/**
- * Shorten the dynamic param in order to make it smaller when transmitted to the browser.
- */
-function getShortDynamicParamType(
-  type: DynamicParamTypes
-): DynamicParamTypesShort {
-  switch (type) {
-    case 'catchall':
-      return 'c'
-    case 'optional-catchall':
-      return 'oc'
-    case 'dynamic':
-      return 'd'
-    default:
-      throw new Error('Unknown dynamic param type')
-  }
-}
-
-/**
- * Parse dynamic route segment to type of parameter
- */
-function getSegmentParam(segment: string): {
-  param: string
-  type: DynamicParamTypes
-} | null {
-  if (segment.startsWith('[[...') && segment.endsWith(']]')) {
-    return {
-      type: 'optional-catchall',
-      param: segment.slice(5, -2),
-    }
-  }
-
-  if (segment.startsWith('[...') && segment.endsWith(']')) {
-    return {
-      type: 'catchall',
-      param: segment.slice(4, -1),
-    }
-  }
-
-  if (segment.startsWith('[') && segment.endsWith(']')) {
-    return {
-      type: 'dynamic',
-      param: segment.slice(1, -1),
-    }
-  }
-
-  return null
-}
-
-/**
- * Get inline <link> tags based on server CSS manifest. Only used when rendering to HTML.
- */
-function getCssInlinedLinkTags(
-  clientReferenceManifest: ClientReferenceManifest,
-  serverCSSManifest: ClientCSSReferenceManifest,
-  filePath: string,
-  serverCSSForEntries: string[],
-  injectedCSS: Set<string>,
-  collectNewCSSImports?: boolean
-): string[] {
-  const layoutOrPageCssModules = serverCSSManifest.cssImports[filePath]
-
-  const filePathWithoutExt = filePath.replace(/\.[^.]+$/, '')
-  const cssFilesForEntry = new Set(
-    clientReferenceManifest.cssFiles?.[filePathWithoutExt] || []
-  )
-
-  if (!layoutOrPageCssModules || !cssFilesForEntry.size) {
-    return []
-  }
-  const chunks = new Set<string>()
-
-  for (const mod of layoutOrPageCssModules) {
-    // We only include the CSS if it's a global CSS, or it is used by this
-    // entrypoint.
-    if (
-      serverCSSForEntries.includes(mod) ||
-      !/\.module\.(css|sass|scss)$/.test(mod)
-    ) {
-      // If the CSS is already injected by a parent layer, we don't need
-      // to inject it again.
-      if (!injectedCSS.has(mod)) {
-        const modData =
-          clientReferenceManifest.clientModules[
-            getClientReferenceModuleKey(mod, '')
-          ]
-        if (modData) {
-          for (const chunk of modData.chunks) {
-            // If the current entry in the final tree-shaked bundle has that CSS
-            // chunk, it means that it's actually used. We should include it.
-            if (cssFilesForEntry.has(chunk)) {
-              chunks.add(chunk)
-              // This might be a new layout, and to make it more efficient and
-              // not introducing another loop, we mutate the set directly.
-              if (collectNewCSSImports) {
-                injectedCSS.add(mod)
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-
-  return [...chunks]
-}
-
-function getServerCSSForEntries(
-  serverCSSManifest: ClientCSSReferenceManifest,
-  entries: string[]
-) {
-  const css = []
-  for (const entry of entries) {
-    const entryName = entry.replace(/\.[^.]+$/, '')
-    if (
-      serverCSSManifest.cssModules &&
-      serverCSSManifest.cssModules[entryName]
-    ) {
-      css.push(...serverCSSManifest.cssModules[entryName])
-    }
-  }
-  return css
-}
-
-/**
- * Get inline <link rel="preload" as="font"> tags based on server CSS manifest and next/font manifest. Only used when rendering to HTML.
- */
-function getPreloadedFontFilesInlineLinkTags(
-  serverCSSManifest: ClientCSSReferenceManifest,
-  nextFontManifest: NextFontManifest | undefined,
-  serverCSSForEntries: string[],
-  filePath: string | undefined,
-  injectedFontPreloadTags: Set<string>
-): string[] | null {
-  if (!nextFontManifest || !filePath) {
-    return null
-  }
-  const layoutOrPageCss = serverCSSManifest.cssImports[filePath]
-
-  if (!layoutOrPageCss) {
-    return null
-  }
-
-  const fontFiles = new Set<string>()
-  let foundFontUsage = false
-
-  for (const css of layoutOrPageCss) {
-    // We only include the CSS if it is used by this entrypoint.
-    if (serverCSSForEntries.includes(css)) {
-      const preloadedFontFiles = nextFontManifest.app[css]
-      if (preloadedFontFiles) {
-        foundFontUsage = true
-        for (const fontFile of preloadedFontFiles) {
-          if (!injectedFontPreloadTags.has(fontFile)) {
-            fontFiles.add(fontFile)
-            injectedFontPreloadTags.add(fontFile)
-          }
-        }
-      }
-    }
-  }
-
-  // If we find an entry in the manifest but it's empty, add a preconnect tag by returning null.
-  // Only render a preconnect tag if we previously didn't preload any fonts.
-  if (
-    !foundFontUsage ||
-    (fontFiles.size === 0 && injectedFontPreloadTags.size > 0)
-  ) {
-    return null
-  }
-
-  // Sorting to make order deterministic
-  return [...fontFiles].sort()
-}
-
-function getScriptNonceFromHeader(cspHeaderValue: string): string | undefined {
-  const directives = cspHeaderValue
-    // Directives are split by ';'.
-    .split(';')
-    .map((directive) => directive.trim())
-
-  // First try to find the directive for the 'script-src', otherwise try to
-  // fallback to the 'default-src'.
-  const directive =
-    directives.find((dir) => dir.startsWith('script-src')) ||
-    directives.find((dir) => dir.startsWith('default-src'))
-
-  // If no directive could be found, then we're done.
-  if (!directive) {
-    return
-  }
-
-  // Extract the nonce from the directive
-  const nonce = directive
-    .split(' ')
-    // Remove the 'strict-src'/'default-src' string, this can't be the nonce.
-    .slice(1)
-    .map((source) => source.trim())
-    // Find the first source with the 'nonce-' prefix.
-    .find(
-      (source) =>
-        source.startsWith("'nonce-") &&
-        source.length > 8 &&
-        source.endsWith("'")
-    )
-    // Grab the nonce by trimming the 'nonce-' prefix.
-    ?.slice(7, -1)
-
-  // If we could't find the nonce, then we're done.
-  if (!nonce) {
-    return
-  }
-
-  // Don't accept the nonce value if it contains HTML escape characters.
-  // Technically, the spec requires a base64'd value, but this is just an
-  // extra layer.
-  if (ESCAPE_REGEX.test(nonce)) {
-    throw new Error(
-      'Nonce value from Content-Security-Policy contained HTML escape characters.\nLearn more: https://nextjs.org/docs/messages/nonce-contained-invalid-characters'
-    )
-  }
-
-  return nonce
-}
-
-async function renderToString(element: React.ReactElement) {
-  return getTracer().trace(AppRenderSpan.renderToString, async () => {
-    const renderStream = await ReactDOMServer.renderToReadableStream(element)
-    await renderStream.allReady
-    return streamToString(renderStream)
-  })
-}
-
-function parseAndValidateFlightRouterState(
-  stateHeader: string | string[] | undefined
-): FlightRouterState | undefined {
-  if (typeof stateHeader === 'undefined') {
-    return undefined
-  }
-  if (Array.isArray(stateHeader)) {
-    throw new Error(
-      'Multiple router state headers were sent. This is not allowed.'
-    )
-  }
-
-  // We limit the size of the router state header to ~40kb. This is to prevent
-  // a malicious user from sending a very large header and slowing down the
-  // resolving of the router state.
-  // This is around 2,000 nested or parallel route segment states:
-  // '{"children":["",{}]}'.length === 20.
-  if (stateHeader.length > 20 * 2000) {
-    throw new Error('The router state header was too large.')
-  }
-
-  try {
-    return flightRouterStateSchema.parse(JSON.parse(stateHeader))
-  } catch {
-    throw new Error('The router state header was sent but could not be parsed.')
-  }
-}
-
-function validateURL(url: string | undefined): string {
-  if (!url) {
-    throw new Error('Invalid request URL')
-  }
-  try {
-    new URL(url, 'http://n')
-    return url
-  } catch {
-    throw new Error('Invalid request URL')
-  }
-}
+export const isEdgeRuntime = process.env.NEXT_RUNTIME === 'edge'
 
 export async function renderToHTMLOrFlight(
   req: IncomingMessage,
