+++ packages/next/src/server/app-render/interop-default.ts
@@ -1,0 +1,6 @@
+/**
+ * Interop between "export default" and "module.exports".
+ */
+export function interopDefault(mod: any) {
+  return mod.default || mod
+}
